--story=1006222 --user=龙儒锋 profile所属系统添加全局引用、profile引用全局参数 https://www.tapd.cn/54247054/s/1223419

+++ src/main/java/codedriver/module/autoexec/api/profile/AutoexecProfileSaveApi.java
@@ -69,6 +69,9 @@ public class AutoexecProfileSaveApi extends PrivateApiComponentBase {
         if (paramProfileId != null && autoexecProfileMapper.checkProfileIsExists(paramProfileId) == 0) {
             throw new AutoexecProfileIsNotFoundException(paramProfileId);
         }
+        if (autoexecProfileMapper.checkProfileNameIsRepeats(profileVo) > 0) {
+            throw new AutoexecProfileNameRepeatsException(profileVo.getName());
+        }
 
         //删除profile和tool、script的关系
         autoexecProfileMapper.deleteProfileOperationByProfileId(paramProfileId);
